requirements.txt
python-dotenv
weaviate-client
pandas